LPS-54878 Moves code from EditLayoutPrototypeAction to Portlet, no logic changes
